This is the nginx config file.
This is an important file:

/etc/nginx/sites-available/default

A small error can cause some super strange results:

1) I had an issue where href was set to "/angular". If you hit refresh on the angular app, the index page for scottlarrribeau.com would load
instead of scottlarribeau.com/angular. Also, the scottlarribeau.com/index page would not have any graphic assets.

This turned out because of a bug in the nginx default file:

This is what the file should have contained:

        location /angular/ {
                root /var/www/html/angular;
                try_files $uri $uri/ /index.html =404;
        }

This is what was actually there:
        location /angular/ {
                root /var/www/html
                try_files $uri $uri/ /index.html =404;
        }

Now the problem was that on a refresh of the url, instead of scottlarribeau.com/angular/index.com loading, we were getting
scottlarribeau.com/index.com. Since nginix was currently pointing to the angular directory, when it loaded the index file from /,
there were none of the assets that index.html from / was expecting.

Cool bug,